Add support for LNS protocol options
This commit removes the `lns_address` and `protocol` fields from
Core.Route and reimplements them as a single field populated by a
struct.  The struct may be one of:

  - Core.HttpRoamingLns
  - Core.GwmpLns
  - Core.HeliumRouterLns

A schema, DB.Lns, is added to store these structs in a normalized
format in a new `route_lns` database table.